+++ managers/flow_builder.py
@@ -73,6 +73,8 @@ class FlowBuilder:
                 break
 
         if not new_int_flow_tbl_0_tcp:
+            if not evc["active"]:
+                return []
             raise FlowsNotFound(evc["id"])
 
         utils.set_instructions_from_actions(new_int_flow_tbl_0_tcp)
@@ -222,6 +224,8 @@ class FlowBuilder:
             new_int_flow_tbl_0_tcp = copy.deepcopy(flow)
 
             if not new_int_flow_tbl_0_tcp:
+                if not evc["active"]:
+                    return []
                 raise FlowsNotFound(evc["id"])
 
             utils.set_new_cookie(new_int_flow_tbl_0_tcp)