+++ CHANGELOG.rst
@@ -13,6 +13,10 @@ Added
 - Handled ``kytos/mef_eline.(redeployed_link_down|redeployed_link_up)`` event.
 - Handled ``kytos/mef_eline.error_redeploy_link_down`` event.
 
+Changed
+=======
+- Only raise ``FlowsNotFound`` when an EVC is active and flows aren't found. Update status and status_reason accordingly too when installing flows.
+
 [2023.2.0] - 2024-02-16
 ***********************